requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pandas==2.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uuid7==0.1.0
uvicorn==0.38.0
twilio==8.10.0  
requests==2.31.0  
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
firebase-admin==6.2.0
# passlib[bcrypt]==1.7.4
python-multipart==0.0.6
passlib==1.7.4
bcrypt==3.2.0

